requirements.txt
asgiref==3.8.1
colorama==0.4.6
crispy-bootstrap5==2024.2
Django==5.0.6
django-appconf==1.0.6
django-crispy-forms==2.2
django_compressor==4.5
djangorestframework==3.15.2
factory-boy==3.3.0
Faker==26.0.0
iniconfig==2.0.0
mysqlclient==2.2.4
packaging==24.1
pillow==10.3.0
pluggy==1.5.0
psycopg==3.1.19
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
rcssmin==1.1.2
rjsmin==1.2.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
